stonepotapi/requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
django==4.1.5
gunicorn==20.1.0
djangorestframework==3.12.4
django-cors-headers==3.10.1
pytz==2021.3
pymongo==3.12.1